Prototype turn-based interraction
At this stage you can start a game via an invite and play roughly alternate
turns. The match data is the .sgf backup and any turn clobbers any ongoing
game. This is meant as a rough proof of concept and is likely to change
substantially now that the mechanics of both game center and littlego are better
understood.

There are many TODOs and hacks.